feat: add function that truncates domain b/c of rapid base level rise
This function takes the domain and a new base level (shore_z), then computes the closest node to that, and resizes the domain accordingly. It then also resizes the other vectors on the domain to pick the subset of the channel that remains subaerial.

In the future, may want to do some linear interpolation or something, but this works for now.